requirements.txt
tensorflow==2.19.0
pandas 
numpy 
scikit-learn
tensorboard
matplotlib
streamlit
scikeras
altair>=4.0,<5.0